+++ Core.hpp
@@ -11,12 +11,12 @@ SC_MODULE(Core)
 {
 public:
 	SC_HAS_PROCESS(Core);
-	Core(sc_module_name name, int& sv_num, const int sv_len, sc_event *e_ready, sc_event *e_next, int& lambda, int& target, deque<double> &data);
+	Core(sc_module_name name, int& sv_num, const int sv_len, sc_event *e_ready, sc_event *e_next, double& lambda, int& target, deque<double> &data);
 	void proc();
 protected:
 	sc_event *e_next;
 	sc_event *e_ready;
-	int &lambda;
+	double &lambda;
 	int &target;
 	deque <double> &data;
 	deque <double> y;
